+++ src/shuriken/meta.clj
@@ -1,0 +1,6 @@
+(ns shuriken.meta)
+
+(defmacro without-meta
+  "Sets meta of x to nil."
+  [x]
+  `(with-meta ~x nil))