+++ test/shuriken/namespace_test.clj
@@ -1,4 +1,4 @@
-(ns shuriken.core-test
+(ns shuriken.namespace-test
   (:require [clojure.test :refer :all]
             [shuriken.core :refer [fully-qualify]]
             [shuriken.virtual-test-namespace :refer [a-var] :as virtual])
@@ -6,10 +6,10 @@
 
 (deftest test-fully-qualify
   (let [local-var :local]
-    (binding [*ns* (find-ns 'shuriken.core-test)]
+    (binding [*ns* (find-ns 'shuriken.namespace-test)]
       (are [sym expected] (= (fully-qualify *ns* sym) expected)
            'Object              'java.lang.Object
-           'fully-qualify       'shuriken.core/fully-qualify
+           'fully-qualify       'shuriken.namespace/fully-qualify
            'local-var           'local-var
            'a-var               'shuriken.virtual-test-namespace/a-var
            'virtual/another-var 'shuriken.virtual-test-namespace/another-var