+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.3.0] - 2016-11-20
+### Added
+  - associative: flatten-keys, deflatten-keys, deep-merge
+  - meta: without-meta
+
 ## [0.2.0] - 2016-11-20
 ### Added
-  - fully-qualify
+  - namespace: fully-qualify
 
 ## [0.1.0] - 2016-11-20
 ### Added
-  - predicate composers : `and?`, `or?` & `not?`.
+  - predicate composers: `and?`, `or?` & `not?`.